requirements.txt
python-dotenv==1.0.1
requests==2.31.0
openai==1.14.2
charset-normalizer==3.3.2
urllib3
idna==3.6
certifi==2024.2.2
tqdm==4.66.2
sniffio==1.3.1
distro==1.9.0
anyio==4.3.0
httpx==0.27.0
pydantic==2.6.4
typing-extensions==4.9.0
exceptiongroup==1.2.0
python-dotenv==1.0.1
requests==2.31.0
openai==1.14.2
langchain==0.1.12
psycopg2==2.9.9
unstructured==0.12.6
tiktoken==0.6.0
pgvector==0.2.5
langchain-openai==0.0.8

